simple form for voting
The value of each radio button is the associated question choice’s ID. The name of each radio button is "choice". That means, when somebody selects one of the radio buttons and submits the form, it’ll send the POST data choice=# where # is the ID of the selected choice.

We set the form’s action to {% url 'polls:vote' question.id %}, and we set method="post".